Uproszczenie PhotoControllera - wprowadzenie interfejsu IAsyncRepository.cs do IPhotoRepository.cs - Drobne poprawki w plikach cshtml, - Refaktoryzacja kodu w controllerze, serwisie oraz repozytorium, - Usunięcie zbędnych elementów - Rozdzielenie logiki repozytorium od controllera,

+++ Controllers/PhotosController.cs
@@ -12,9 +12,6 @@ namespace Luxa.Controllers
     public class PhotosController(IUserService userService, ApplicationDbContext context, UserManager<UserModel> userManager, IWebHostEnvironment hostEnvironment, IPhotoService photoService, ICommentService commentService) : Controller
     {
         private readonly IPhotoService _photoService = photoService;
-
-        private readonly ApplicationDbContext _context = context; //wstrzykiwanie kontekstu
-
         private readonly UserManager<UserModel> _userManager = userManager;
         private readonly IWebHostEnvironment _hostEnvironment = hostEnvironment;
         private readonly IUserService _userService = userService;
@@ -24,87 +21,41 @@ namespace Luxa.Controllers
         public async Task<IActionResult> Index()
             //Nie powinno się z tego co wiem przekazywać UserModela do widoku ale inaczej bez tworzenia ViewModelu
             //Bez Include się nie wyświetla ale trzeba to przerobić albo na vm albo pobawić się ViewBagami
-            => View(await _context.Photo.Include(m => m.Owner).ToListAsync());
+            // => View(await _context.Photo.Include(m => m.Owner).ToListAsync());
+            => View(await _photoService.GetPhotosWithOwner());
 
         [Authorize]
-        public IActionResult DownloadImage(int id)
+        public async Task<IActionResult> DownloadImage(int id)
         {
-            var photo = _photoService.GetImageById(id);
+            var photo = await _photoService.GetImageByIdAsync(id);
+            if (photo == null || string.IsNullOrEmpty(photo.Name))
+                return RedirectToAction("Error", "Home");
             string filePath = Path.Combine(_hostEnvironment.WebRootPath, "Image/" + photo.Name);
-            byte[] fileBytes = System.IO.File.ReadAllBytes(filePath);
-
-            if (photo == null)
-            {
-                return null;
-            }
+            byte[] fileBytes = await System.IO.File.ReadAllBytesAsync(filePath);
             string extension = Path.GetExtension(photo.Name);
-            string contentType = GetContentType(extension);
+            string contentType = _photoService.GetContentType(extension);
             return File(fileBytes, contentType, photo.Name);
         }
-
-        private string GetContentType(string extension)
-        {
-            return extension.ToLower() switch
-            {
-                ".jpg" => "image/jpeg",
-                ".jpeg" => "image/jpeg",
-                ".png" => "image/png",
-                ".gif" => "image/gif",
-                _ => "application/octet-stream",
-            };
-        }
-
-        /*public async Task<IActionResult> Index()
-        {
-            var photos = await _photoService.GetAllImagesAsync();
-            return View(photos);
-        }*/
-
-
-
-
-        // GET: Photos/Details/5
+        
         [Authorize]
         public async Task<IActionResult> Details(int id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var photo = await _context.Photo
-                .Include(m => m.Owner)
-                .FirstOrDefaultAsync(m => m.Id == id);
-
+            var photo = await _photoService.GetOneWithOwner(id);
             if (photo == null)
             {
                 return NotFound();
             }
-            // Pobierz komentarze dla tego zdjęcia
             var comments = await _commentService.GetCommentsForPhoto(id);
-            // Przekaż komentarze do widoku
             ViewData["Comments"] = comments;
-
-            // Przekazanie id zdjęcia do widoku
             ViewBag.PhotoId = id;
-
-
             return View(photo);
         }
 
         // GET: Photos/Create
         [Authorize]
         public IActionResult Create()
-        {
+            => _userManager.GetUserId(User) == null ? View("Views/Home/Index.cshtml") : View();
 
-            if (_userManager.GetUserId(User) == null)
-            {
-                return View("Views/Home/Index.cshtml");
-            }
-
-            return View();
-            //filtr privacy
-        }
 
         // POST: Photos/Create
         // To protect from overposting attacks, enable the specific properties you want to bind to.
@@ -149,10 +100,7 @@ namespace Luxa.Controllers
         public async Task<IActionResult> Edit(int id, Photo photo)
         {
             if (id != photo.Id)
-            {
-                return NotFound();
-            }
-
+                return RedirectToAction("Error", "Home");
             if (!ModelState.IsValid)
             {
                 try
@@ -165,30 +113,17 @@ namespace Luxa.Controllers
                     {
                         return NotFound();
                     }
-                    else
-                    {
-                        throw;
-                    }
+                    throw;
                 }
                 return RedirectToAction("Index", "Home");
             }
-            else
+            var errors = (from state in ModelState
+                from error in state.Value.Errors
+                select error.ErrorMessage).ToList();
+            // Wyświetl błędy w konsoli (lub zrób coś innego z błędami)
+            foreach (var error in errors)
             {
-                var errors = new List<string>();
-
-                foreach (var state in ModelState)
-                {
-                    foreach (var error in state.Value.Errors)
-                    {
-                        errors.Add(error.ErrorMessage);
-                    }
-                }
-
-                // Wyświetl błędy w konsoli (lub zrób coś innego z błędami)
-                foreach (var error in errors)
-                {
-                    Console.WriteLine(error);
-                }
+                Console.WriteLine(error);
             }
             return View(photo);
         }
@@ -198,17 +133,10 @@ namespace Luxa.Controllers
         public async Task<IActionResult> Delete(int? id)
         {
             if (id == null)
-            {
                 return NotFound();
-            }
-
-            var photo = await _context.Photo
-                .FirstOrDefaultAsync(m => m.Id == id);
+            var photo = await _photoService.GetImageByIdAsync((int)id);
             if (photo == null)
-            {
                 return NotFound();
-            }
-
             return View(photo);
         }
 
@@ -218,25 +146,15 @@ namespace Luxa.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var photo = await _context.Photo.FindAsync(id);
-            //delete image from wwwroot/image
+            var photo = await _photoService.GetImageByIdAsync(id);
+            if (photo == null)
+                return RedirectToAction("Error", "Home");
             var imagePath = Path.Combine(_hostEnvironment.WebRootPath, "image", photo.Name);
             if (System.IO.File.Exists(imagePath))
-            {
                 System.IO.File.Delete(imagePath);
-            }
-
-            //delete record
-            _context.Photo.Remove(photo);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            return (await _photoService.Delete(photo)) ? RedirectToAction(nameof(Index)) : RedirectToAction("Error","Home");
         }
-
-        private bool PhotoExists(int id)
-        {
-            return _context.Photo.Any(e => e.Id == id);
-        }
-
+        
 
         [HttpPost]
         [Authorize]
@@ -248,7 +166,7 @@ namespace Luxa.Controllers
                 return false;
             var likedPhotos = await _photoService.GetLikedPhotos(user);
             return (!_photoService.IsPhotoLiked(idPhoto, likedPhotos))
-                ? _photoService.LikePhoto(idPhoto, user)
+                ? await _photoService.LikePhoto(idPhoto, user)
                 : _photoService.UnlikePhoto(idPhoto, user);
         }
     }
